Cleanup CustomList object
