+++ Shared/Configuration/Watchlist.xcdatamodeld/Watchlist.xcdatamodel/contents
@@ -3,9 +3,7 @@
     <entity name="CustomList" representedClassName="CustomList" syncable="YES" codeGenerationType="class">
         <attribute name="creationDate" optional="YES" attributeType="Date" usesScalarValueType="NO"/>
         <attribute name="id" optional="YES" attributeType="UUID" usesScalarValueType="NO"/>
-        <attribute name="lastItemImage" optional="YES" attributeType="URI"/>
         <attribute name="notes" optional="YES" attributeType="String"/>
-        <attribute name="shared" attributeType="Boolean" defaultValueString="NO" usesScalarValueType="YES"/>
         <attribute name="title" attributeType="String" defaultValueString=""/>
         <attribute name="updatedDate" optional="YES" attributeType="Date" usesScalarValueType="NO"/>
         <relationship name="items" optional="YES" toMany="YES" deletionRule="Nullify" destinationEntity="WatchlistItem" inverseName="list" inverseEntity="WatchlistItem"/>
