+++ Shared/Localization/pt-BR.lproj/Localizable.strings
@@ -513,3 +513,4 @@
 "allCustomWatchListFilters" = "Tudo";
 "moviesCustomWatchListFilters" = "Filmes";
 "showsCustomWatchListFilters" = "Séries";
+"%lld items" = "%lld items";
