+++ Shared/View/CustomList/NewCustomListView.swift
@@ -127,7 +127,6 @@ struct NewCustomListView: View {
         list.title = title
         list.creationDate = Date()
         list.updatedDate = Date()
-        list.shared = false
         list.notes = note
         list.items = itemsToAdd as NSSet
         print(list as Any)
@@ -154,121 +153,5 @@ struct NewCustomListView_Previews: PreviewProvider {
 #else
         EmptyView()
 #endif
-    }
-}
-
-struct EditCustomList: View {
-#if os(macOS)
-    @Binding var isPresentingNewList: Bool
-#endif
-    @State var list: CustomList
-    @State private var title = ""
-    @State private var note = ""
-    @State private var shareList = false
-    @State private var hasUnsavedChanges = false
-    @State private var disableSaveButton = true
-    @Binding var showListSelection: Bool
-    @State private var itemsToRemove = Set<String>()
-    var body: some View {
-        Form {
-            Section {
-                TextField("listName", text: $title)
-                TextField("listDescription", text: $note)
-            } header: {
-                Label("listBasicHeader", systemImage: "pencil")
-            }
-            
-            Section {
-                if list.itemsArray.isEmpty {
-                    
-                } else {
-                    List {
-                        ForEach(list.itemsArray, id: \.notificationID) { item in
-                            HStack {
-                                Image(systemName: itemsToRemove.contains(item.notificationID) ? "minus.circle.fill" : "circle")
-                                    .foregroundColor(itemsToRemove.contains(item.notificationID) ? .red : nil)
-                                WebImage(url: item.image)
-                                    .resizable()
-                                    .placeholder {
-                                        ZStack {
-                                            Rectangle().fill(.gray.gradient)
-                                            Image(systemName: item.itemMedia == .movie ? "film" : "tv")
-                                        }
-                                    }
-                                    .aspectRatio(contentMode: .fill)
-                                    .frame(width: 70, height: 50)
-                                    .cornerRadius(6)
-                                    .overlay {
-                                        if itemsToRemove.contains(item.notificationID) {
-                                            ZStack {
-                                                Rectangle().fill(.black.opacity(0.4))
-                                            }
-                                            .cornerRadius(6)
-                                        }
-                                    }
-                                VStack(alignment: .leading) {
-                                    Text(item.itemTitle)
-                                        .lineLimit(1)
-                                        .foregroundColor(itemsToRemove.contains(item.notificationID) ? .secondary : nil)
-                                    Text(item.itemMedia.title)
-                                        .font(.caption)
-                                        .foregroundColor(.secondary)
-                                }
-                            }
-                            .onTapGesture {
-                                if itemsToRemove.contains(item.notificationID) {
-                                    itemsToRemove.remove(item.notificationID)
-                                } else {
-                                    itemsToRemove.insert(item.notificationID)
-                                }
-                            }
-                        }
-                    }
-                }
-            } header: {
-                Label("editListRemoveItems", systemImage: "rectangle.on.rectangle.slash")
-            }
-        }
-#if os(macOS)
-        .formStyle(.grouped)
-#endif
-        .onAppear {
-            title = list.itemTitle
-            note = list.notes ?? ""
-            shareList = list.shared
-        }
-        .onChange(of: title, perform: { newValue in
-            if newValue != list.itemTitle {
-                disableSaveButton = false
-            }
-        })
-        .onChange(of: note, perform: { newValue in
-            if newValue != list.notes {
-                disableSaveButton = false
-            }
-        })
-        .onAppear {
-#if os(macOS)
-            isPresentingNewList = true
-#endif
-        }
-        .onDisappear {
-#if os(macOS)
-            isPresentingNewList = false
-#endif
-        }
-        .toolbar {
-            Button("Save") {
-                PersistenceController.shared.updateListInformation(list: list,
-                                                                   title: title,
-                                                                   description: note)
-                showListSelection = false
-            }
-            .disabled(disableSaveButton)
-#if os(macOS)
-            .buttonStyle(.link)
-#endif
-        }
-        .navigationTitle(list.itemTitle)
     }
 }
