+++ Shared/View/Navigation/SearchView.swift
@@ -62,12 +62,6 @@ struct SearchView: View {
             .navigationDestination(for: [Person].self) { items in
                 DetailedPeopleList(items: items)
             }
-            .navigationDestination(for: ProductionCompany.self) { item in
-                //CompanyDetails(company: item)
-            }
-            .navigationDestination(for: [ProductionCompany].self) { item in
-                //CompaniesListView(companies: item)
-            }
             .searchable(text: $viewModel.query,
                         placement: .navigationBarDrawer(displayMode: .always),
                         prompt: Text("Movies, Shows, People"))
